Add files via upload
SQL-based Supply Chain Analysis and visualization of Supply chain analysis using Power BI for visualizing the findings, to enhance the understanding of the analyzed data, Power BI is employed as a visualization Tool. The project leverages these capabilities to create an interactive dashboard that allows users to explore data and drill down into specific details.